+++ CHANGELOG.md
@@ -42,6 +42,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2026-03-03
 ### Added
-- Initial release of the StreamFlow prototype.
+- Initial release of the Slalom prototype.
 - Basic support for POSIX and HTTP adapters.
 - Composition Root (Bootstrap) architecture.